PM / devfreq: Add MSM BIMC bwmon support for bw_hwmon governor
The BIMC bwmon device supports monitoring read/write traffic from each BIMC
master port. It also has the capability to raise an IRQ when the traffic
count exceeds a programmable threshold. This allows for it to be used with
the bw_hwmon governor to scale the BW requests from each BIMC master.

Change-Id: Ie8a1471226411e23954ed556292186a5a864ddc1
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
[avajid@codeaurora.org: made minor styling changes]
Signed-off-by: Amir Vajid <avajid@codeaurora.org>
[Diaz1401: sync with https://git.codelinaro.org/clo/la/kernel/msm-5.4/-/commit/716a02830df859fb50a1ab082292177199c0cbf2]
Signed-off-by: Diaz1401 <reagor8161@outlook.com>
Conflicts:
	drivers/devfreq/Kconfig
	drivers/devfreq/Makefile
	drivers/devfreq/bimc-bwmon.c