mptcp: avoid meta window shrink impacting subflow
Dropping data due to out-of-window on a subflow in MPTCP can occur
due to the window shrinking at meta level. This can happen for example
due to reinjected data sent on another subflow filling the window. The
latter semi-frequently occurs due to penalization while initially
growing the congestion window.  For the subflow, dropping the packet is
seen (wrongly) as a loss, impacting the congestion control.

Keep track of the highest announced right edge of the receive window and
validate incoming skb with the corresponding window size of this max
right edge, even when the window has shrunk and the incoming skb would
not fit in the currently announced window.
Basically, if we told the sender that it could send the segment at some
point, we shouldn't refuse the segment when it arrives.

This bug was observed impacting the congestion window growth of an
end-to-end test where one link had a much larger rtt (500ms versus
20ms). The slow flow would sometimes incur a loss very early due to
data arriving on a zero window. This slows down the initial growth of
the cwnd substantially.

Notes about back- or forward-porting this patch follow:
- rcv_right_edge should be updated anytime rcv_wnd or rcv_wup get updated.
  If rcv_right_edge might not have been initialized yet, it has to be
  explicitly set. If it has been initialized, a call to
  tcp_update_rcv_right_edge is required.
- when validating if an incoming packet should be accepted,
  tcp_receive_window_no_shrink should be used. In other cases,
  tcp_receive_window can be used.

Because tcp_receive_window_no_shrink will validate against the most
permissive of the current receive window and the
right-most-right-edge-based receive window, the patch should even be
robust if rcv_right_edge has not been perfectly updated, and even if it
has not been initialised. Of course, this could lead to accepting
packets that were sent out of window, and should be avoided.

Fixes: Zero-day Bug
Suggested-by: Christoph Paasch <cpaasch@apple.com>
Co-developed-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Tim Froidcoeur <tim.froidcoeur@tessares.net>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit d8e3bd88da5fbee8e8ccfc9b3f4aa56e11ee06ce)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Diaz1401 <reagor8161@outlook.com>