mptcp: disable BH before increase stats from rcu-critical sections
Similar to
commit 874432801016 ("mptcp: Disable bottom-half before accessing rcu-critical sections"),
since
commit 5413d1babe8f ("net: do not block BH while processing socket backlog"),
packets can be processed with bottom-half enabled.

This breaks our assumption across the code that everything in TCP is
being done with BH disabled. Thus, we need to increase the stats with
BH except when it is explicitly disabled before.

This was mainly visible with a v5.4 kernel, see the issue below.

We can also remove MPTCP_INC_STATS_BH() which was already unused since
the introduction of the similar fix.

Fixes: 669b5db7d4dc ("Merge tag 'v4.7' into mptcp_trunk")
Closes: https://github.com/multipath-tcp/mptcp/issues/429
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit dbc37a8471e380fa1230838a7417b3867e33baa2)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Diaz1401 <reagor8161@outlook.com>