mptcp: remember backup flag sent in the join SYN+ACK
When the stack sets the backup flag in the SYN+ACK for a join, it stores
the information in request_sock. Whenever the real sock is allocated
when the 3-way handshake completes, the information needs to be copied.

The scheduler considers a subflow as backup if either side sets the
backups flags.

Without this change, if the client is not setting the backup flag but
the server is, then the scheduler would ignore the flags.

Fixes: Zero-day bug
Signed-off-by: Benjamin Hesmans <benjamin.hesmans@tessares.net>
Acked-by: Christoph Paasch <cpaasch@apple.com>
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit 27118cdd871865ddeba510fcea6ed4d400a9983c)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
(cherry picked from commit a8a1e9490a5015b70f8c2d77b1dc9aa264f1aab1)
Signed-off-by: Matthieu Baerts <matthieu.baerts@tessares.net>
Signed-off-by: Diaz1401 <reagor8161@outlook.com>