kernel: add a config for building with PGO
GCC can make use of PGO (Profile-Guided Optimization) to enhance the
efficiency and performance of the kernel. This technique involves
analyzing the code to gather data about frequently executed parts and
frequently called functions. The compiler then utilizes this data to
optimize the most vital parts of the code, which leads to significant
performance gains, especially for larger and more complex areas of
the kernel.

Change-Id: Iaf5cd778445f7f7447c01513e5d2e7f3c7a6c8fa
Signed-off-by: Vaisakh Murali <mvaisakh@statixos.com>
Signed-off-by: Diaz1401 <reagor8161@outlook.com>