Makefile: rework PGO flags
* import some flags enabled by '-fprofile-use'
'-fprofile-reorder-functions', Function reordering based on profile instrumentation collects first time of execution of a function and orders these functions in ascending order.
'-ftracer', Perform tail duplication to enlarge superblock size. This transformation simplifies the control flow of the function allowing other optimizations to do a better job.

* '-fprofile-correction'
When this option is specified, GCC uses heuristics to correct or smooth out such inconsistencies

* '-fprofile-partial-training'
With -fprofile-partial-training profile feedback will be ignored for all functions not executed during the train run leading them to be optimized as if they were compiled without profile feedback.
This leads to better performance when train run is not representative but also leads to significantly bigger code.

Signed-off-by: Diaz1401 <diaznura@outlook.com>